refactor: restructure the starter files
