+++ src/components/Blog/bio.jsx
@@ -9,7 +9,7 @@ import React from 'react';
 import { useStaticQuery, graphql } from 'gatsby';
 import Image from 'gatsby-image';
 
-import { rhythm } from '../utils/typography';
+import { rhythm } from '../../utils/typography';
 
 const Bio = () => {
   const data = useStaticQuery(graphql`
