+++ src/pages/blog/index.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { Link, graphql } from 'gatsby';
 
-import Bio from '../components/bio';
-import Layout from '../components/layout';
-import SEO from '../components/seo';
-import { rhythm } from '../utils/typography';
+import Bio from '../../components/Blog/bio';
+import Layout from '../../components/Blog/layout';
+import SEO from '../../components/Blog/seo';
+import { rhythm } from '../../utils/typography';
 
 const BlogIndex = ({ data, location }) => {
   const siteTitle = data.site.siteMetadata.title;
