+++ src/templates/blog-post.jsx
@@ -1,12 +1,15 @@
 import React from 'react';
 import { Link, graphql } from 'gatsby';
 
-import Bio from '../components/bio';
-import Layout from '../components/layout';
-import SEO from '../components/seo';
+import Bio from '../components/Blog/bio';
+import Layout from '../components/Blog/layout';
+import SEO from '../components/Blog/seo';
 import { rhythm, scale } from '../utils/typography';
 
-const BlogPostTemplate = ({ data, pageContext, location }) => {
+const BlogPostTemplate = (forExample) => {
+  const { location } = forExample;
+  const { pageContext } = forExample;
+  const { data } = forExample;
   const post = data.markdownRemark;
   const siteTitle = data.site.siteMetadata.title;
   const { previous, next } = pageContext;
